requirements.txt
fastapi~=0.115.5
uvicorn~=0.32.1
pydantic~=2.10.2
SQLAlchemy~=2.0.36
passlib~=1.7.4
PyYAML~=6.0.2